k8s: use self implementation of reSync
As the k8s reSync functionality only sends "adds" events whenever it
refetches all data from k8s, this implementation will make sure the
objects are not re-added based on their Equalness function
implementation.

In this commit the old reSync behavior is not changed as the Equalness
functions lacks a proper implementation.

Also, in k8s watcher, if a delete event is missed, the watcher will
never be aware of this deletion as the reSync doesn't keep any state of
the watched objects.

In this commit the delete event behavior is also unchanged as the
missingFunc lacks a proper implementation.

This commit only introduces the framework that will allow to write the
dedicated Equalness and MissingFunc for each object being watched.

After proper implementation of those dedicated functions it is expected
that on every reSync there will be a AddFunc event for objects that had
relevant modifications to be processed.

Signed-off-by: André Martins <andre@cilium.io>
Signed-off-by: André Martins <aanm@Andres-MacBook-Air.local>